+++ genometools/geo/generate_sample_sheet.py
@@ -23,6 +23,7 @@
 from __future__ import (absolute_import, division,
                         print_function, unicode_literals)
 from builtins import *
+import six
 
 import sys
 import os
@@ -51,13 +52,20 @@ def get_argument_parser():
     g = parser.add_argument_group('Input and output files')
 
     g.add_argument(
-        '-s', '--series-matrix-file', type=str, required=True,
+        '-s', '--series-matrix-file', type=cli.str_type, required=True,
         metavar=cli.file_mv, help='The GEO series matrix file.'
     )
 
     g.add_argument(
-        '-o', '--output-file', type=str, required=True,
+        '-o', '--output-file', type=cli.str_type,
+        required=True,
         metavar=cli.file_mv, help='The output file.'
+    )
+
+    g.add_argument(
+        '-e', '--encoding', type=cli.str_type,
+        metavar=cli.str_mv, default='UTF-8',
+        help='The encoding of the series matrix file. [UTF-8]'
     )
 
     cli.add_reporting_args(parser)
@@ -65,7 +73,7 @@ def get_argument_parser():
     return parser
 
 
-def read_series_matrix(path):
+def read_series_matrix(path, encoding):
     """Read the series matrix."""
     assert isinstance(path, str)
 
@@ -73,7 +81,7 @@ def read_series_matrix(path):
     titles = None
     celfile_urls = None
     with misc.smart_open_read(path, mode='rb', try_gzip=True) as fh:
-        reader = csv.reader(fh, dialect='excel-tab')
+        reader = csv.reader(fh, dialect='excel-tab', encoding=encoding)
         for l in reader:
             if not l:
                 continue
@@ -89,6 +97,23 @@ def read_series_matrix(path):
     return accessions, titles, celfile_urls
 
 
+def write_sample_sheet(path, accessions, names, celfile_urls, sel=None):
+    """Write the sample sheet."""
+    with open(path, 'wb') as ofh:
+        writer = csv.writer(ofh, dialect='excel-tab',
+                            lineterminator=os.linesep,
+                            quoting=csv.QUOTE_NONE)
+        # write header
+        writer.writerow(['Accession', 'Name', 'CEL file', 'CEL file URL'])
+        n = len(names)
+        if sel is None:
+            sel = range(n)
+        for i in sel:
+            cf = celfile_urls[i].split('/')[-1]
+            # row = [accessions[i], names[i], cf, celfile_urls[i]]
+            writer.writerow([accessions[i], names[i], cf, celfile_urls[i]])
+
+
 def main(args=None):
     """Script entry point."""
 
@@ -96,8 +121,12 @@ def main(args=None):
         parser = get_argument_parser()
         args = parser.parse_args()
 
+    #series_matrix_file = newstr(args.series_matrix_file, 'utf-8')
+    #output_file = newstr(args.output_file, 'utf-8')
+    #encoding = newstr(args.encoding, 'utf-8')
     series_matrix_file = args.series_matrix_file
     output_file = args.output_file
+    encoding = args.encoding
 
     # log_file = args.log_file
     # quiet = args.quiet
@@ -106,7 +135,8 @@ def main(args=None):
     # logger = misc.get_logger(log_file = log_file, quiet = quiet,
     #        verbose = verbose)
 
-    accessions, titles, celfile_urls = read_series_matrix(series_matrix_file)
+    accessions, titles, celfile_urls = read_series_matrix(
+        series_matrix_file, encoding=encoding)
     write_sample_sheet(output_file, accessions, titles, celfile_urls)
 
     return 0